Revert GCC 6/7 changes for now.
Higher GCC version is causing cellular network problems like:
 - random signal loss which needs one or two reboots to fix,
 - intermittent disconnection while using LTE/3G data

So stick to GCC 4.9 for now and revert the following changes:
 - Revert "drivers: prima: Fix all GCC 6/7 warnings"
 - Revert "athene: Fix some more GCC-6/7 warnings"
 - Revert "Fix all GCC 7.x warnings"
 - Revert "Fix all GCC 6.x warnings"
 - Revert "Makefile: Fix device not booting with GCC 7.x and above"

+++ drivers/usb/gadget/f_fs.c
@@ -1552,7 +1552,7 @@ static void functionfs_unbind(struct ffs_data *ffs)
 static int ffs_epfiles_create(struct ffs_data *ffs)
 {
 	struct ffs_epfile *epfile, *epfiles;
-	short i, count;
+	unsigned i, count;
 
 	ENTER();
